Rename services
The concept of an "Unauthenticated" service isn't quite useful
