+++ packages/notifi-axios-adapter/lib/index.ts
@@ -1,1 +1,1 @@
-export * from './AxiosUnauthenticatedService';
+export * from './AxiosService';
