+++ packages/notifi-core/lib/services/NotifiService.ts
@@ -1,0 +1,3 @@
+import { LogInFromDaoService } from '.';
+
+export type NotifiService = LogInFromDaoService;
